requirements.txt
discord.py>=2.0.0,<3
python-dotenv==0.20.0
psycopg2
